API root should return JSON
It's kind of jarring to send a request, expect JSON and then get hit
witha wall of text because somebody set up a redirect to an HTML
site.

When users now hit the base `/api` route they will get a JSON
response with links to the docs and other relevant material.

Updates #10