Fix First/Last Agg Error
The API was a little too naive in the way that it applied pagination and
ordering. In the case of calling `first` or `last` value aggregates for
observations, the defaults applied would break the query.

To solve this, I changed the underlying query to only grab the latest
record rather than using TimescaleDB's functions (which don't really do
what I wanted to do anyway). I also updated the ordering and pagination
plugs to have keyword lists of assigned key/values on the connection act
as short circuits.

Fixes #15