Massive Updates -- V2
This is a pretty seismic change to the code. There are minimal changes
to the public API and routes.

The goal of this was to slim down the API footprint but cutting out
filters that were literally never used, add rate limiting, add
websockets and provide some clean landing pages and information.